Updated DAO (needs design adjustments and tons of testing! but bulk of it is complete)
