+++ src/data_access/test_files/GenerateRoutineDataAccessObject.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.testing;
 
 import entity.Exercise;
 import entity.ExerciseFactory;
